fix odd mixin crash, 1.15.2
