+++ gradle.properties
@@ -5,15 +5,15 @@ org.gradle.jvmargs=-Xmx1G
 # check these on https://fabricmc.net/use
 minecraft_version=1.17.1
 yarn_mappings=1.17.1+build.61
-loader_version=0.11.7
+loader_version=0.12.4
 
 # Mod Properties
-mod_version=1.15.1
+mod_version=1.15.2
 maven_group=draylar
 archives_base_name=identity
 
 # Dependencies
 # currently not on the main fabric site, check on the maven: https://maven.fabricmc.net/net/fabricmc/fabric-api/fabric-api
-fabric_version=0.40.1+1.17
+fabric_version=0.41.3+1.17
 cardinal_version=3.1.1
 pal_version=1.3.0-nightly.1.17-rc1
