+++ src/main/java/draylar/identity/mixin/LivingEntityMixin.java
@@ -2,6 +2,7 @@ package draylar.identity.mixin;
 
 import draylar.identity.Identity;
 import draylar.identity.api.IdentityGranting;
+import draylar.identity.impl.NearbySongAccessor;
 import draylar.identity.registry.Components;
 import draylar.identity.registry.EntityTags;
 import net.fabricmc.api.EnvType;
@@ -13,14 +14,18 @@ import net.minecraft.entity.damage.DamageSource;
 import net.minecraft.entity.effect.StatusEffect;
 import net.minecraft.entity.effect.StatusEffectInstance;
 import net.minecraft.entity.effect.StatusEffects;
+import net.minecraft.entity.mob.SpiderEntity;
 import net.minecraft.entity.mob.WaterCreatureEntity;
 import net.minecraft.entity.passive.BatEntity;
+import net.minecraft.entity.passive.DolphinEntity;
 import net.minecraft.entity.player.PlayerEntity;
 import net.minecraft.fluid.Fluid;
+import net.minecraft.tag.FluidTags;
 import net.minecraft.util.math.BlockPos;
 import net.minecraft.world.World;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.Shadow;
+import org.spongepowered.asm.mixin.Unique;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.ModifyVariable;
@@ -29,7 +34,7 @@ import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
 
 @Mixin(LivingEntity.class)
-public abstract class LivingEntityMixin extends Entity {
+public abstract class LivingEntityMixin extends Entity implements NearbySongAccessor {
 
     @Shadow
     protected abstract int getNextAirOnLand(int air);
@@ -186,28 +191,62 @@ public abstract class LivingEntityMixin extends Entity {
 
     @Inject(method = "hurtByWater", at = @At("HEAD"), cancellable = true)
     protected void identity_hurtByWater(CallbackInfoReturnable<Boolean> cir) {
-        // NO-OP
+        if((LivingEntity) (Object) this instanceof PlayerEntity player) {
+            LivingEntity entity = Components.CURRENT_IDENTITY.get(player).getIdentity();
+
+            if (entity != null) {
+                cir.setReturnValue(entity.hurtByWater());
+            }
+        }
     }
 
     @Inject(method = "canBreatheInWater", at = @At("HEAD"), cancellable = true)
     protected void identity_canBreatheInWater(CallbackInfoReturnable<Boolean> cir) {
-        // NO-OP
+        if((LivingEntity) (Object) this instanceof PlayerEntity player) {
+            LivingEntity entity = Components.CURRENT_IDENTITY.get(player).getIdentity();
+
+            if (entity != null) {
+                cir.setReturnValue(entity.canBreatheInWater() || entity instanceof DolphinEntity || EntityTags.UNDROWNABLE.contains(entity.getType()));
+            }
+        }
     }
+
+    @Unique
+    private boolean nearbySongPlaying = false;
 
     @Environment(EnvType.CLIENT)
     @Inject(method = "setNearbySongPlaying", at = @At("RETURN"))
     protected void identity_setNearbySongPlaying(BlockPos songPosition, boolean playing, CallbackInfo ci) {
-        // NO-OP
+        if((LivingEntity) (Object) this instanceof PlayerEntity player) {
+            nearbySongPlaying = playing;
+        }
+    }
+
+    @Override
+    public boolean identity_isNearbySongPlaying() {
+        return nearbySongPlaying;
     }
 
     @Inject(method = "isUndead", at = @At("HEAD"), cancellable = true)
     protected void identity_isUndead(CallbackInfoReturnable<Boolean> cir) {
-        // NO-OP
+        if((LivingEntity) (Object) this instanceof PlayerEntity player) {
+            LivingEntity identity = Components.CURRENT_IDENTITY.get(player).getIdentity();
+
+            if (identity != null) {
+                cir.setReturnValue(identity.isUndead());
+            }
+        }
     }
 
     @Inject(method = "canWalkOnFluid", at = @At("HEAD"), cancellable = true)
     protected void identity_canWalkOnFluid(Fluid fluid, CallbackInfoReturnable<Boolean> cir) {
-        // NO-OP
+        if((LivingEntity) (Object) this instanceof PlayerEntity player) {
+            LivingEntity identity = Components.CURRENT_IDENTITY.get(player).getIdentity();
+
+            if (identity != null && EntityTags.LAVA_WALKING.contains(identity.getType()) && fluid.isIn(FluidTags.LAVA)) {
+                cir.setReturnValue(true);
+            }
+        }
     }
 
     @Inject(
@@ -216,6 +255,12 @@ public abstract class LivingEntityMixin extends Entity {
             cancellable = true
     )
     protected void identity_allowSpiderClimbing(CallbackInfoReturnable<Boolean> cir) {
-        // NO-OP
+        if((LivingEntity) (Object) this instanceof PlayerEntity player) {
+            LivingEntity identity = Components.CURRENT_IDENTITY.get(player).getIdentity();
+
+            if (identity instanceof SpiderEntity) {
+                cir.setReturnValue(this.horizontalCollision);
+            }
+        }
     }
 }
